fix: db migrations applied
