+++ backend/alembic/versions/3dcb55a66ca1_initial_migration_with_all_models.py
@@ -1,0 +1,34 @@
+"""Initial migration with all models
+
+Revision ID: 3dcb55a66ca1
+Revises: 
+Create Date: 2025-07-28 14:23:39.796372
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = '3dcb55a66ca1'
+down_revision: Union[str, None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.add_column('medicines', sa.Column('doctor_id', sa.Integer(), nullable=True))
+    op.create_index(op.f('ix_medicines_doctor_id'), 'medicines', ['doctor_id'], unique=False)
+    op.create_foreign_key(None, 'medicines', 'doctors', ['doctor_id'], ['id'])
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_constraint(None, 'medicines', type_='foreignkey')
+    op.drop_index(op.f('ix_medicines_doctor_id'), table_name='medicines')
+    op.drop_column('medicines', 'doctor_id')
+    # ### end Alembic commands ###
